refactor files for react hooks evm and cosmos
