+++ packages/valence-domain-clients-react/package.json
@@ -22,10 +22,11 @@
     "@valence-protocol/valence-domain-clients-core": "workspace:*"
   },
   "peerDependencies": {
-    "react": "^18.0.0"
+    "react": "^18.0.0",
+    "wagmi": "^2.16.0"
   },
   "devDependencies": {
-    "typescript": "^5.0.0",
-    "@types/react": "^18.0.0"
+    "@types/react": "^18.0.0",
+    "typescript": "^5.0.0"
   }
-} 
+}
