+++ packages/valence-domain-clients-react/src/common/hooks/useValenceDomainConfig.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { ValenceDomainClientsConfigContext } from "@/common";
+
+export function useValenceDomainConfig() {
+    const config = useContext(ValenceDomainClientsConfigContext);
+    if (!config) throw new Error('useValenceDomainConfig must be used within a ValenceDomainClientsProvider');
+    return config;
+  }
