+++ packages/valence-domain-clients-react/src/cosmos/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useValenceCosmosConfig';
+export * from './useSigningCosmosClient';
