+++ packages/valence-domain-clients-react/src/evm/context/index.ts
@@ -1,0 +1,1 @@
+export * from './WagmiProvider';
