+++ packages/valence-domain-clients-react/src/evm/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useValenceEvmConfig';
+export * from './useSigningEvmClient';
