--- packages/valence-domain-clients-react/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./useSigningEvmClient";
-export * from "./useSigningCosmosClient";
